Moved CameraMarker under components
